Create src
This program will help the user to determine which mortgage of 2 possibilities is best. The formula for computing
the monthly payment for a loan is:
Payment = (L*Rate/12)/(1-(1+Rate/12) –m) Where:
 L is the amount of the loan
 M is the duration of the loan in months (120 would be 10 years, for example)
 Rate is the annual interest rate (ex: 6.25%)
Your program will ask the user for the amount of the loan along with 2 loan terms (duration and rate). It will
calculate for you the monthly payment for each loan along with the total amount to be paid over the life of the
loan. It will also tell you the lowest monthly payment between the two loans. Here are 2 sample runs:
 Please enter the values requested for the 2 loans you wish to compare
 What is the amount of the loan? 25000

 First set of terms
 Annual interest rate (Ex. 3.5%): 7.5%
 Number of months: 20

 Second set of terms
 Annual interest rate (Ex. 3.5%): 5%
 Number of months: 12

 The monthly payment for loan 1 is $1,333.65
 The monthly payment for loan 2 is $2,140.19
 Therefore your minimum payment would be $1333.65
 With the first loan, you would be paying a total of $26,672.99
 With the second loan, you would be paying a total of $25,682.24
 Please enter the values requested for the 2 loans you wish to compare
 What is the amount of the loan? 100000

 First set of terms
 Annual interest rate (Ex. 3.5%): 5.25%
 Number of months: 240

 Second set of terms
 Annual interest rate (Ex. 3.5%): 7%
 Number of months: 120

 The monthly payment for loan 1 is $673.84
 The monthly payment for loan 2 is $1,161.08
 Therefore your minimum payment would be $673.84
 With the first loan, you would be paying a total of $161,722.60
 With the second loan, you would be paying a total of $139,330.18
Please notice that the values are entered on the same line as the prompts, the dollar format of the output (use
NumberFormat methods for this), and that the rate must be entered as a string (because of the %) and then
changed to a double. Recall also that 5.25% is 0.0525 when used as a decimal in the formula.
As usual, for full credit get this lab checked by Tuesday at 8:00 and upload your code. Recall that you can get
partial credit for unchecked code.